chore: made useSwitchTheme a exportable function
