+++ app/components/home.tsx
@@ -17,6 +17,7 @@ import { Chat } from "./chat";
 
 import dynamic from "next/dynamic";
 import { ErrorBoundary } from "./error";
+import { useSwitchTheme } from "../function/useSwitchTheme";
 
 export function Loading(props: { noLogo?: boolean }) {
   return (
@@ -34,39 +35,6 @@ const Settings = dynamic(async () => (await import("./settings")).Settings, {
 const ChatList = dynamic(async () => (await import("./chat-list")).ChatList, {
   loading: () => <Loading noLogo />,
 });
-
-function useSwitchTheme() {
-  const config = useChatStore((state) => state.config);
-
-  useEffect(() => {
-    document.body.classList.remove("light");
-    document.body.classList.remove("dark");
-
-    if (config.theme === "dark") {
-      document.body.classList.add("dark");
-    } else if (config.theme === "light") {
-      document.body.classList.add("light");
-    }
-
-    const metaDescriptionDark = document.querySelector(
-      'meta[name="theme-color"][media]',
-    );
-    const metaDescriptionLight = document.querySelector(
-      'meta[name="theme-color"]:not([media])',
-    );
-
-    if (config.theme === "auto") {
-      metaDescriptionDark?.setAttribute("content", "#151515");
-      metaDescriptionLight?.setAttribute("content", "#fafafa");
-    } else {
-      const themeColor = getComputedStyle(document.body)
-        .getPropertyValue("--theme-color")
-        .trim();
-      metaDescriptionDark?.setAttribute("content", themeColor);
-      metaDescriptionLight?.setAttribute("content", themeColor);
-    }
-  }, [config.theme]);
-}
 
 const useHasHydrated = () => {
   const [hasHydrated, setHasHydrated] = useState<boolean>(false);
